Remove utilities for reading properties as invoking Java is too slow
