+++ java/configuration/src/main/java/sleeper/configuration/properties/InstanceProperty.java
@@ -16,12 +16,4 @@
 package sleeper.configuration.properties;
 
 public interface InstanceProperty extends SleeperProperty {
-
-    static InstanceProperty fromName(String propertyName) {
-        InstanceProperty userDefined = UserDefinedInstanceProperty.fromName(propertyName);
-        if (userDefined != null) {
-            return userDefined;
-        }
-        return SystemDefinedInstanceProperty.fromName(propertyName);
-    }
 }
