+++ scripts/utility/downloadConfig.sh
@@ -30,4 +30,4 @@ SCRIPTS_DIR=$(cd "$(dirname "$0")" && cd "../" && pwd)
 GENERATED_DIR=${SCRIPTS_DIR}/generated
 
 rm -r "${GENERATED_DIR:?}"/*
-java -cp "${SCRIPTS_DIR}"/jars/clients-*-utility.jar sleeper.status.config.DownloadConfig "$INSTANCE_ID" "$GENERATED_DIR"
+java -cp "${SCRIPTS_DIR}"/jars/clients-*-utility.jar sleeper.status.update.DownloadConfig "$INSTANCE_ID" "$GENERATED_DIR"
